Update JavaFXPlugin to lazy-load and configure task
Motivation
----------

When using the JavaFX Gradle plugin in conjunction with Gradle Enterprise, the
build scan identified that this plugin was eagerly loading a task into the task
graph when it did not need to. This slows down large builds with hundreds of
modules and tasks. Since Gradle 5.0 the concept of task configuration
avoidance [1] was added. This helps speed up builds by only registering tasks
when they are needed. This commit updates the JavaFX plugin to use task
configuration avoidance.

Modifications
-------------

Moving to task configuration avoidance is fairly straightforward. There is a
guide in the documentation [1]. The instance of the create() method was changed
to the corresponding register() method.

Result
------

The result is a Gradle plugin that leverages task configuration avoidance
improving build times.

[1] https://docs.gradle.org/current/userguide/task_configuration_avoidance.html